+++ addOns/paramdigger/src/main/java/org/zaproxy/addon/paramdigger/gui/CustomColumn.java
@@ -3,7 +3,7 @@
  *
  * ZAP is an HTTP/HTTPS proxy for assessing web application security.
  *
- * Copyright 2022 The ZAP Development Team
+ * Copyright 2023 The ZAP Development Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,10 +17,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.paramdigger;
+package org.zaproxy.addon.paramdigger.gui;
 
-import java.util.EventListener;
+@SuppressWarnings("serial")
+public abstract class CustomColumn<T> {
+    Class<?> columnClass;
+    String name;
 
-public interface ParamDiggerResultEventListener extends EventListener {
-    void notifyResult(ParamGuessResultEvent event);
+    public CustomColumn(Class<?> columnClass, String name) {
+        this.columnClass = columnClass;
+        this.name = name;
+    }
+
+    public Class<?> getColumnClass() {
+        return columnClass;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public abstract Object getValue(T model);
+
+    public Object getPrototypeValue() {
+        return null;
+    }
 }