+++ addOns/paramdigger/src/main/java/org/zaproxy/addon/paramdigger/gui/ParamDiggerHistoryTableModel.java
@@ -35,6 +35,11 @@ public class ParamDiggerHistoryTableModel extends DefaultHistoryReferencesTableM
                     Column.URL,
                     Column.STATUS_CODE,
                     Column.STATUS_REASON,
+                    Column.SIZE_MESSAGE,
+                    Column.SIZE_REQUEST_HEADER,
+                    Column.SIZE_REQUEST_BODY,
+                    Column.SIZE_RESPONSE_HEADER,
+                    Column.SIZE_RESPONSE_BODY,
                     Column.RTT,
                 });
     }