+++ addOns/paramdigger/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update minimum ZAP version to 2.12.0.
 - Maintenance changes.
 - Default number of threads to 2 * processor count.
+- Change panel designs to allow message selection.
 
 ## [0.1.0] - 2022-08-22
 ### Changed